requirement.txt
phidata 
python-dotenv
yfinance 
duckduckgo-search
fastapi
uvicorn
groq
packaging
